requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
distribute==0.6.24
greenlet==0.4.2
motor==0.3
pymongo==2.7.1
pyzmq==14.3.1
tornado==3.2.2
wsgiref==0.1.2
